Se actualiza el json y se agregan errores

+++ entrega_4/routers/productos.js
@@ -1,6 +1,5 @@
 const express = require('express');
-//const { Router } = express;
-//router = express.Router()
+
 const productosRouter = express.Router();
 
 const Contenedor = require('../../entrega_3/contenedor');
@@ -16,15 +15,18 @@ productosRouter.get('/', async (req, res) =>{
 productosRouter.get('/:id', async (req, res) =>{
     idProducto = Number(req.params.id)
     const productoSeleccionado = await productosContenedor.getById(idProducto)
-    res.send({
-        data: productoSeleccionado
-    })
+    if (!productoSeleccionado){
+        res.send({ error : 'producto no encontrado' })
+    }else{
+        res.send({
+            data: productoSeleccionado
+        })
+    }
 })
 
 productosRouter.post('/', async (req, res) =>{
     const newProducto = req.body; 
     const idProductoNuevo = await productosContenedor.save(newProducto);
-
     res.send({
         message : 'success',
         data: {
@@ -55,10 +57,13 @@ productosRouter.put('/:id', async (req, res) =>{
 
 productosRouter.delete('/:id', async (req, res) =>{
     idProducto = Number(req.params.id)
-    console.log(idProducto)
-    await productosContenedor.deleteById(idProducto);
+    if (!idProducto){
+        res.send({ error : 'producto no encontrado' })
+    }else {
+        await productosContenedor.deleteById(idProducto);
+        res.send({ message : 'Producto Eliminado de Forma Correcta' })
+    }
 
-    res.send({ message : 'Producto Eliminado de Forma Correcta' })
 })
 
 module.exports = productosRouter;
